requirements/local.txt
-r base.txt
beautifulsoup4
fabric
gitpython
ipdb
hgapi
IPy
psycopg2
pytest-django
pyyaml
walkdir
